[REFACTOR-021] extract operations-types, AttachmentsSection, LogPaymentForm, PendingPaymentsSection, ItemForm; fix import paths
