+++ app/admin/operations/components/PaymentsTab.tsx
@@ -1,23 +1,13 @@
 'use client'
 
 import { useState } from 'react'
-import { useMutation, useQueryClient } from '@tanstack/react-query'
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 import { formatDate, formatCurrency } from '@/lib/format'
 import { Drawer } from '@/components/ui/drawer'
-import {
-  Select,
-  SelectTrigger,
-  SelectValue,
-  SelectContent,
-  SelectItem,
-  SelectLabel,
-  SelectSeparator,
-  SelectGroup,
-} from '@/components/ui/select'
-import type { Trip } from './TripsTab'
-import type { PayableItem } from './ItemsTab'
-import { useQuery } from '@tanstack/react-query'
 import { t } from '@/lib/i18n'
+import type { Trip, PayableItem, MembersResponse, MemberProfile } from './operations-types'
+import { LogPaymentForm } from './LogPaymentForm'
+import { PendingPaymentsSection } from './PendingPaymentsSection'
 
 // ── Types ────────────────────────────────────────────────
 
@@ -39,12 +29,8 @@ type Payment = {
   payable_items: { title: string; item_type: string; currency: string } | null
 }
 
-export type MembersResponse = {
-  los_members: { profile: { id: string; first_name: string; last_name: string; abo_number: string | null } | null }[]
-  manual_members_no_abo: { id: string; first_name: string; last_name: string; upline_abo_number: string | null }[]
-}
-
-type MemberProfile = { id: string; first_name: string; last_name: string; abo_number: string | null }
+// Re-export for page.tsx backwards compat (page still imports MembersResponse from here)
+export type { MembersResponse } from './operations-types'
 
 // ── Helpers ──────────────────────────────────────────────
 
@@ -65,15 +51,6 @@ export function PaymentsTab({ trips, membersData }: { trips: Trip[]; membersData
   const [drawerOpen, setDrawerOpen] = useState(false)
   const [statusFilter, setStatusFilter] = useState<'all' | 'pending' | 'approved' | 'rejected'>('all')
   const [reviewNotes, setReviewNotes] = useState<Record<string, string>>({})
-
-  const [entity, setEntity] = useState('')
-  const [profileId, setProfileId] = useState('')
-  const [amount, setAmount] = useState('')
-  const [currency, setCurrency] = useState('EUR')
-  const [txDate, setTxDate] = useState(new Date().toISOString().split('T')[0])
-  const [method, setMethod] = useState('')
-  const [payStatus, setPayStatus] = useState<'approved' | 'pending'>('approved')
-  const [note, setNote] = useState('')
   const [payError, setPayError] = useState<string | null>(null)
 
   const { data: items = [] } = useQuery<PayableItem[]>({
@@ -124,14 +101,6 @@ export function PaymentsTab({ trips, membersData }: { trips: Trip[]; membersData
     onSuccess: () => {
       qc.invalidateQueries({ queryKey: ['admin-payments'] })
       setDrawerOpen(false)
-      setEntity('')
-      setProfileId('')
-      setAmount('')
-      setCurrency('EUR')
-      setTxDate(new Date().toISOString().split('T')[0])
-      setMethod('')
-      setPayStatus('approved')
-      setNote('')
       setPayError(null)
     },
     onError: (e: Error) => setPayError(e.message),
@@ -147,31 +116,12 @@ export function PaymentsTab({ trips, membersData }: { trips: Trip[]; membersData
     onSuccess: () => qc.invalidateQueries({ queryKey: ['admin-payments'] }),
   })
 
-  function handleLog() {
-    if (!entity || !profileId || !amount) return
-    const [type, id] = entity.split('::')
-    const body: Record<string, unknown> = {
-      profile_id: profileId,
-      amount: Number(amount),
-      currency,
-      transaction_date: txDate,
-      payment_method: method || null,
-      note: note || null,
-      admin_status: payStatus,
-    }
-    if (type === 'trip') body.trip_id = id
-    else body.payable_item_id = id
-    logMutation.mutate(body)
-  }
-
   const STATUS_FILTERS = [
     { key: 'all' as const, labelKey: 'admin.operations.payments.filter.all' as const },
     { key: 'pending' as const, labelKey: 'admin.operations.payments.filter.pending' as const },
     { key: 'approved' as const, labelKey: 'admin.operations.payments.filter.approved' as const },
     { key: 'rejected' as const, labelKey: 'admin.operations.payments.filter.rejected' as const },
   ]
-
-  const activeItems = items.filter(i => i.is_active)
 
   return (
     <section className="space-y-6">
@@ -185,56 +135,14 @@ export function PaymentsTab({ trips, membersData }: { trips: Trip[]; membersData
         </button>
       </div>
 
-      {pendingSubmissions.length > 0 && (
-        <div>
-          <p className="text-xs font-semibold tracking-widest uppercase mb-3" style={{ color: 'var(--text-secondary)' }}>
-            {t('admin.operations.payments.pendingTitle', 'en').replace('{{count}}', String(pendingSubmissions.length))}
-          </p>
-          <div className="rounded-2xl border overflow-hidden" style={{ backgroundColor: 'var(--bg-card)', borderColor: 'var(--border-default)' }}>
-            {pendingSubmissions.map((p, i) => (
-              <div key={p.id} className="px-5 py-4 space-y-3"
-                style={{ borderTop: i > 0 ? '1px solid var(--border-default)' : 'none' }}>
-                <div className="flex items-start justify-between gap-4">
-                  <div>
-                    <p className="text-sm font-semibold" style={{ color: 'var(--text-primary)' }}>
-                      {p.profiles?.first_name} {p.profiles?.last_name}
-                      {p.profiles?.abo_number && <span className="font-mono font-normal ml-2 text-xs opacity-60">{p.profiles.abo_number}</span>}
-                    </p>
-                    <p className="text-xs mt-0.5" style={{ color: 'var(--text-secondary)' }}>
-                      {p.trips?.title ?? p.payable_items?.title ?? '—'}
-                      {' · '}{formatDate(p.transaction_date)}
-                      {' · '}{formatCurrency(p.amount, p.currency)}
-                      {p.payment_method && ` · ${p.payment_method}`}
-                    </p>
-                    {p.note && <p className="text-xs mt-0.5 italic" style={{ color: 'var(--text-secondary)' }}>{p.note}</p>}
-                    {p.proof_url && <a href={p.proof_url} target="_blank" rel="noopener noreferrer" className="text-xs hover:underline" style={{ color: 'var(--brand-teal)' }}>{t('admin.operations.payments.viewProof', 'en')}</a>}
-                  </div>
-                  <span className="text-[10px] font-semibold px-2 py-0.5 rounded-full flex-shrink-0" style={{ backgroundColor: '#f2cc8f33', color: '#7a5c00' }}>{t('admin.operations.payments.badge.pending', 'en')}</span>
-                </div>
-                <div className="flex items-center gap-3">
-                  <input
-                    value={reviewNotes[p.id] ?? ''}
-                    onChange={e => setReviewNotes(n => ({ ...n, [p.id]: e.target.value }))}
-                    placeholder={t('admin.operations.payments.placeholder.adminNote', 'en')}
-                    className="flex-1 border rounded-xl px-3 py-2 text-xs"
-                    style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }}
-                  />
-                  <button
-                    onClick={() => reviewMutation.mutate({ id: p.id, admin_status: 'approved', admin_note: reviewNotes[p.id] || null })}
-                    disabled={reviewMutation.isPending}
-                    className="px-4 py-2 rounded-xl text-xs font-semibold text-white disabled:opacity-40 hover:opacity-90 transition-opacity flex-shrink-0"
-                    style={{ backgroundColor: '#2d6a4f' }}>{t('admin.operations.payments.btn.approve', 'en')}</button>
-                  <button
-                    onClick={() => reviewMutation.mutate({ id: p.id, admin_status: 'rejected', admin_note: reviewNotes[p.id] || null })}
-                    disabled={reviewMutation.isPending}
-                    className="px-4 py-2 rounded-xl text-xs font-semibold text-white disabled:opacity-40 hover:opacity-90 transition-opacity flex-shrink-0"
-                    style={{ backgroundColor: 'var(--brand-crimson)' }}>{t('admin.operations.payments.btn.deny', 'en')}</button>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
+      <PendingPaymentsSection
+        payments={pendingSubmissions}
+        reviewNotes={reviewNotes}
+        setReviewNotes={setReviewNotes}
+        onApprove={(id, note) => reviewMutation.mutate({ id, admin_status: 'approved', admin_note: note })}
+        onReject={(id, note) => reviewMutation.mutate({ id, admin_status: 'rejected', admin_note: note })}
+        isPending={reviewMutation.isPending}
+      />
 
       <div className="flex gap-2 flex-wrap">
         {STATUS_FILTERS.map(f => (
@@ -283,103 +191,18 @@ export function PaymentsTab({ trips, membersData }: { trips: Trip[]; membersData
         </div>
       )}
 
-      <Drawer open={drawerOpen} onClose={() => setDrawerOpen(false)} title={t('admin.operations.payments.drawer.title', 'en')}>
-        <div className="space-y-4">
-          <div>
-            <label className="text-xs mb-1 block" style={{ color: 'var(--text-secondary)' }}>{t('admin.operations.payments.lbl.entity', 'en')}</label>
-            <Select value={entity} onValueChange={setEntity}>
-              <SelectTrigger><SelectValue placeholder={t('admin.operations.payments.placeholder.entity', 'en')} /></SelectTrigger>
-              <SelectContent>
-                {trips.length > 0 && (
-                  <SelectGroup>
-                    <SelectLabel>{t('admin.operations.payments.group.trips', 'en')}</SelectLabel>
-                    {trips.map(t2 => <SelectItem key={t2.id} value={`trip::${t2.id}`}>{t2.title}</SelectItem>)}
-                  </SelectGroup>
-                )}
-                {trips.length > 0 && activeItems.length > 0 && <SelectSeparator />}
-                {activeItems.length > 0 && (
-                  <SelectGroup>
-                    <SelectLabel>{t('admin.operations.payments.group.items', 'en')}</SelectLabel>
-                    {activeItems.map(it => <SelectItem key={it.id} value={`item::${it.id}`}>{it.title}</SelectItem>)}
-                  </SelectGroup>
-                )}
-              </SelectContent>
-            </Select>
-          </div>
-          <div>
-            <label className="text-xs mb-1 block" style={{ color: 'var(--text-secondary)' }}>{t('admin.operations.payments.lbl.member', 'en')}</label>
-            <Select value={profileId} onValueChange={setProfileId}>
-              <SelectTrigger><SelectValue placeholder={t('admin.operations.payments.placeholder.member', 'en')} /></SelectTrigger>
-              <SelectContent>
-                {allMembers.map(m => (
-                  <SelectItem key={m.id} value={m.id}>
-                    {m.last_name}, {m.first_name}{m.abo_number ? ` · ${m.abo_number}` : ''}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-          <div className="grid grid-cols-2 gap-4">
-            <div>
-              <label className="text-xs mb-1 block" style={{ color: 'var(--text-secondary)' }}>{t('admin.operations.payments.lbl.amount', 'en')}</label>
-              <input type="number" step="0.01" min="0" value={amount} onChange={e => setAmount(e.target.value)}
-                placeholder="0.00" className="w-full border rounded-xl px-3 py-2.5 text-sm"
-                style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }} />
-            </div>
-            <div>
-              <label className="text-xs mb-1 block" style={{ color: 'var(--text-secondary)' }}>{t('admin.operations.payments.lbl.currency', 'en')}</label>
-              <input value={currency} onChange={e => setCurrency(e.target.value)}
-                className="w-full border rounded-xl px-3 py-2.5 text-sm"
-                style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }} />
-            </div>
-          </div>
-          <div className="grid grid-cols-2 gap-4">
-            <div>
-              <label className="text-xs mb-1 block" style={{ color: 'var(--text-secondary)' }}>{t('admin.operations.payments.lbl.date', 'en')}</label>
-              <input type="date" value={txDate} onChange={e => setTxDate(e.target.value)}
-                className="w-full border rounded-xl px-3 py-2.5 text-sm"
-                style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }} />
-            </div>
-            <div>
-              <label className="text-xs mb-1 block" style={{ color: 'var(--text-secondary)' }}>{t('admin.operations.payments.lbl.method', 'en')}</label>
-              <input value={method} onChange={e => setMethod(e.target.value)}
-                placeholder={t('admin.operations.payments.placeholder.method', 'en')}
-                className="w-full border rounded-xl px-3 py-2.5 text-sm"
-                style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }} />
-            </div>
-          </div>
-          <div>
-            <label className="text-xs mb-1 block" style={{ color: 'var(--text-secondary)' }}>{t('admin.operations.payments.lbl.status', 'en')}</label>
-            <Select value={payStatus} onValueChange={val => setPayStatus(val as 'approved' | 'pending')}>
-              <SelectTrigger><SelectValue /></SelectTrigger>
-              <SelectContent>
-                <SelectItem value="approved">{t('admin.operations.payments.status.approved', 'en')}</SelectItem>
-                <SelectItem value="pending">{t('admin.operations.payments.status.pending', 'en')}</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-          <div>
-            <label className="text-xs mb-1 block" style={{ color: 'var(--text-secondary)' }}>
-              {t('admin.operations.payments.lbl.note', 'en')} <span className="opacity-60 font-normal">{t('admin.operations.payments.lbl.noteOptional', 'en')}</span>
-            </label>
-            <input value={note} onChange={e => setNote(e.target.value)}
-              placeholder={t('admin.operations.payments.placeholder.note', 'en')}
-              className="w-full border rounded-xl px-3 py-2.5 text-sm"
-              style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }} />
-          </div>
-          {payError && <p className="text-sm" style={{ color: 'var(--brand-crimson)' }}>{payError}</p>}
-          <div className="flex gap-3 pt-2">
-            <button onClick={handleLog}
-              disabled={logMutation.isPending || !entity || !profileId || !amount || Number(amount) <= 0}
-              className="px-6 py-2.5 rounded-xl text-sm font-semibold text-white disabled:opacity-40 hover:opacity-90 transition-opacity"
-              style={{ backgroundColor: 'var(--brand-crimson)' }}>
-              {logMutation.isPending ? t('admin.operations.payments.btn.saving', 'en') : t('admin.operations.payments.btn.log2', 'en')}
-            </button>
-            <button onClick={() => setDrawerOpen(false)}
-              className="px-6 py-2.5 rounded-xl text-sm font-semibold border transition-colors hover:bg-black/5"
-              style={{ borderColor: 'var(--border-default)', color: 'var(--text-secondary)' }}>{t('admin.operations.payments.btn.cancel', 'en')}</button>
-          </div>
-        </div>
+      <Drawer open={drawerOpen} onClose={() => { setDrawerOpen(false); setPayError(null) }} title={t('admin.operations.payments.drawer.title', 'en')}>
+        <LogPaymentForm
+          trips={trips}
+          items={items}
+          allMembers={allMembers}
+          onSubmit={payload => {
+            if (!payload.profile_id) { setDrawerOpen(false); setPayError(null); return }
+            logMutation.mutate(payload)
+          }}
+          isPending={logMutation.isPending}
+          externalError={payError}
+        />
       </Drawer>
     </section>
   )
